pyproject.toml
[tool.poetry]
name = "lambda-saleor-app"
version = "0.1.0"
description = ""
authors = ["Artur Smęt <artur.smet@mirumee.com>"]
readme = "README.md"
packages = [{include = "lambda_saleor_app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
mangum = "^0.17.0"
structlog = "^22.3.0"
saleor-sdk-python = "^0.0.2"
boto3 = "^1.26.69"
httpx = "^0.23.3"
aws-secrets-cache = "^0.1.0"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.20.0"
black = "^23.1.0"
ipdb = "^0.13.11"
aws-lambda-typing = "^2.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
